requirements.txt
africastalking==1.2.8
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.159
botocore==1.34.159
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
daphne==4.1.2
defusedxml==0.7.1
distlib==0.3.8
Django==5.1
django-admin-extra-buttons==1.5.8
django-allauth==0.63.6
django-appconf==1.0.6
django-auth==0.1.9
django-celery-results==2.5.1
django-compressor==4.4
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-excel==0.0.10
django-filter==24.3
django-haystack==3.3.0
django-log-viewer==1.1.2
django-mathfilters==1.0.0
django-online-users==0.3
django-ratelimit==4.1.0
django-rest-auth==0.9.5
django-simple-history==2.12.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-orjson-renderer==1.7.2
drf-spectacular==0.27.2
et-xmlfile==1.1.0
fcm-django==0.3.7
filelock==3.13.1
gevent==24.2.1
greenlet==3.0.3
hyperlink==21.0.0
identify==2.5.35
idna==3.7
importlib-metadata==3.9.1
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
lml==0.1.0
MarkupSafe==2.1.5
msgpack==1.0.8
nodeenv==1.8.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.0.7
orjson==3.10.7
packaging==24.0
pandas==2.2.0
phonenumbers==8.13.29
pillow==10.4.0
platformdirs==4.2.0
pre-commit==3.8.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyfcm==1.4.9
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
rcssmin==1.1.1
redis==5.0.1
referencing==0.33.0
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.1
rpds-py==0.18.0
s3transfer==0.10.2
schema==0.7.4
sendgrid==6.11.0
service-identity==24.1.0
six==1.15.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
texttable==1.7.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==3.0.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.5
Whoosh==2.7.4
xlrd==2.0.1
zipp==3.4.1
zope.event==5.0
zope.interface==6.2
